requirements.txt
Django==2.1.7
django-filter==2.1.0
django-rest-framework==0.1.0
djangorestframework==3.9.1
djangorestframework-filters==1.0.0.dev0
drf-writable-nested==0.5.1
pytz==2018.9
